+++ src/pages/signup/index.tsx
@@ -1,7 +1,7 @@
 import type { GetStaticProps } from 'next';
 import { getServerTranslations } from '@/lib/i18n/getServerTranslations'
 
-export { default } from '@/components/pages/Signup'
+export { default } from '@/components/pages/Signup/signup'
 
 export const getStaticProps: GetStaticProps = async ({
   locale,
